version up

+++ setup.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #
 # cd /home/pi/
-# wget https://raw.githubusercontent.com/chirimen-oh/chirimen/master/setup.sh
+# wget https://rawgit.com/chirimen-oh/chirimen-raspi3/master/setup.sh
 # ./setup.sh
 #
 # 一時的にスリープを無効
@@ -9,21 +9,39 @@ sudo xset s off
 sudo xset -dpms
 sudo xset s noblank
 # スリープを無効
-sudo sed '1s/$/ consoleblank=0/' /boot/cmdline.txt |\
-    sudo tee /tmp/cmdline && sudo cat /tmp/cmdline |\
-    sudo tee /boot/cmdline.txt && sudo rm -f /tmp/cmdline
+grep 'consoleblank=0' /boot/cmdline.txt
+if [ $? -ge 1 ]; then
+    sudo sed '1s/$/ consoleblank=0/' /boot/cmdline.txt |\
+        sudo tee /tmp/cmdline && sudo cat /tmp/cmdline |\
+        sudo tee /boot/cmdline.txt && sudo rm -f /tmp/cmdline
+fi
 
-echo '@xset s off' | sudo tee -a /etc/xdg/lxsession/LXDE-pi/autostart
-echo '@xset -dpms' | sudo tee -a /etc/xdg/lxsession/LXDE-pi/autostart
-echo '@xset s noblank' | sudo tee -a /etc/xdg/lxsession/LXDE-pi/autostart
+sudo cp /etc/xdg/lxsession/LXDE-pi/autostart /etc/xdg/lxsession/LXDE-pi/autostart.orig
+sudo sh -c "cat << EOF > /etc/xdg/lxsession/LXDE-pi/autostart
+@lxpanel --profile LXDE-pi
+@pcmanfm --desktop --profile LXDE-pi
+@xscreensaver -no-splash
+@xset s off
+@xset -dpms
+@xset s noblank
+@/usr/bin/chromium-browser https://localhost/top --enable-experimental-web-platform-features
+EOF"
 
-# aptをtsukubaに変更
-sudo sed -i '1s/^/#/' /etc/apt/sources.list
-sudo sed -i '1s/^/deb http:\/\/ftp.tsukuba.wide.ad.jp\/Linux\/raspbian\/raspbian\/ buster main contrib non-free rpi\n/' /etc/apt/sources.list
+# aptをmirrorで指定
+sudo sh -c "cat << EOF > /etc/apt/mirrors.txt
+http://ftp.jaist.ac.jp/raspbian/
+http://ftp.tsukuba.wide.ad.jp/Linux/raspbian/raspbian/
+http://ftp.yz.yamagata-u.ac.jp/pub/linux/raspbian/raspbian/
+http://raspbian.raspberrypi.org/raspbian/
+EOF"
+sudo cp /etc/apt/sources.list /etc/apt/sources.list.orig
+sudo sh -c "cat << EOF > /etc/apt/sources.list
+deb mirror+file:/etc/apt/mirrors.txt buster main contrib non-free rpi
+EOF"
 sudo apt-get update
 
 # upgradeを保留に変更
-echo raspberrypi-ui-mods hold | sudo dpkg --set-selections
+sudo apt-mark hold raspberrypi-ui-mods
 # 必要な項目をインストール
 sudo apt-get install at-spi2-core
 
@@ -41,15 +59,10 @@ sudo apt-get -y install ttf-kochi-gothic fonts-noto uim uim-mozc nodejs npm apac
 sudo apt-get -y install ttf-kochi-gothic fonts-noto uim uim-mozc nodejs npm apache2 vim emacs libnss3-tools
 sudo apt-get -y autoremove
 
-# VS code (code-ossのインストール)
-# 最新版だと動かないのでバージョンダウンして固定
-wget -qO - https://packagecloud.io/headmelted/codebuilds/gpgkey | sudo apt-key add -
-sudo apt-get update
-sudo dpkg --configure -a
-curl -s https://code.headmelted.com/installers/apt.sh | sudo bash
-sudo apt-get -y remove code-oss
-sudo apt-get -y install code-oss=1.29.0-1539702286
-sudo apt-mark hold code-oss
+# VS code のインストール
+wget -O /tmp/code.deb 'https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-armhf'
+sudo apt install -y /tmp/code.deb
+
 
 # 日本語設定
 sudo sed 's/#\sen_GB\.UTF-8\sUTF-8/en_GB\.UTF-8 UTF-8/g' /etc/locale.gen |\
@@ -65,7 +78,7 @@ sudo locale-gen ja_JP.UTF-8
 sudo update-locale LANG=ja_JP.UTF-8
 
 # 時間設定
-sudo raspi-config nonint do_change_timezone Japan
+sudo raspi-config nonint do_change_timezone Asia/Tokyo
 
 # キーボード設定
 sudo raspi-config nonint do_configure_keyboard jp
@@ -75,81 +88,191 @@ sudo raspi-config nonint do_wifi_country JP
 
 # node.jsのインストール
 sudo npm install n -g
-sudo n 10.16.0
+sudo n 12.20.0
+PATH=$PATH
 sudo npm i eslint prettier -g
 
-# code-oss extension
-/usr/share/code-oss/bin/code-oss --install-extension dbaeumer.vscode-eslint
-/usr/share/code-oss/bin/code-oss --install-extension esbenp.prettier-vscode
+# VS code extension
+/usr/share/code/bin/code --install-extension dbaeumer.vscode-eslint
+/usr/share/code/bin/code --install-extension esbenp.prettier-vscode
 
-# JSのデフォルトをcode-ossに
-cat << EOS > /home/pi/.config/mimeapps.list
+# JSのデフォルトをVS codeに
+cat << EOF > /home/pi/.config/mimeapps.list
 [Added Associations]
-application/javascript=code-oss.desktop;
+application/javascript=code.desktop;
 
 [Default Applications]
-application/javascript=code-oss.desktop;
-EOS
+application/javascript=code.desktop;
+EOF
 
 # カメラを有効化
 sudo raspi-config nonint do_camera 0
-echo 'options bcm2835-v4l2 gst_v4l2src_is_broken=1' | sudo tee -a /etc/modprobe.d/bcm2835-v4l2.conf
-echo 'bcm2835-v4l2' | sudo tee -a /etc/modules-load.d/modules.conf
+grep 'bcm2835-v4l2' /etc/modprobe.d/bcm2835-v4l2.conf
+if [ $? -ge 1 ]; then
+    echo 'options bcm2835-v4l2 gst_v4l2src_is_broken=1' | sudo tee -a /etc/modprobe.d/bcm2835-v4l2.conf
+fi
+grep 'bcm2835-v4l2' /etc/modules-load.d/modules.conf
+if [ $? -ge 1 ]; then
+    echo 'bcm2835-v4l2' | sudo tee -a /etc/modules-load.d/modules.conf
+fi
 
 # I2Cを有効化
 sudo raspi-config nonint do_i2c 0
 
 # _gc設定
 cd /home/pi/
-wget https://r.chirimen.org/_gc.zip
-unzip ./_gc.zip
+if [ ! -f /home/pi/_gc.zip ]; then
+    wget https://r.chirimen.org/_gc.zip
+fi
+if [ ! -d /home/pi/_gc/ ]; then
+    unzip ./_gc.zip
+fi
 cd /home/pi/_gc/srv
 npm i
 sudo npm i forever -g
 cd /home/pi/
-echo "@reboot sudo -u pi /home/pi/_gc/srv/startup.sh" | crontab
+crontab -l > /tmp/tmp_crontab
+grep '/home/pi/_gc/srv/startup.sh' /tmp/tmp_crontab
+if [ $? = 1 ]; then
+    echo "@reboot sudo -u pi /home/pi/_gc/srv/startup.sh" | crontab
+fi
 ln -s /home/pi/_gc/srv/reset.sh /home/pi/Desktop/reset.sh
 mkdir /home/pi/.config/chromium/
 mkdir /home/pi/.config/chromium/Default/
-mv /home/pi/_gc/bookmark/Bookmarks /home/pi/.config/chromium/Default/Bookmarks
+cp /home/pi/_gc/bookmark/Bookmarks /home/pi/.config/chromium/Default/Bookmarks
 pcmanfm --set-wallpaper /home/pi/_gc/wallpaper/wallpaper-720P.png
 
 
 # gc設定
 chromium-browser &
 cd /home/pi/
-wget https://r.chirimen.org/gc.zip
+if [ ! -f /home/pi/gc.zip ]; then
+    wget https://r.chirimen.org/gc.zip
+fi
 # chromiumの起動待ちダウンロード
-wget https://downloads.arduino.cc/arduino-1.8.6-linuxarm.tar.xz
-unzip ./gc.zip -d /home/pi/Desktop
+if [ ! -f /home/pi/arduino-1.8.13-linuxarm.tar.xz ]; then
+    wget https://downloads.arduino.cc/arduino-1.8.13-linuxarm.tar.xz
+fi
+if [ ! -d /home/pi/Desktop/gc/ ]; then
+    unzip ./gc.zip -d /home/pi/Desktop
+fi
 # chromiumの起動待ち
 sleep 120s
-sudo sed 's/\/var\/www\/html/\/home\/pi\/Desktop\/gc/g' /etc/apache2/sites-available/000-default.conf |\
-    sudo tee /tmp/apache-default && sudo cat /tmp/apache-default |\
-    sudo tee /etc/apache2/sites-available/000-default.conf && sudo rm -f /tmp/apache-default
-sudo sed 's/\/var\/www\//\/home\/pi\/Desktop\/gc/g' /etc/apache2/apache2.conf |\
-    sudo tee /tmp/apache && sudo cat /tmp/apache |\
-    sudo tee /etc/apache2/apache2.conf && sudo rm -f /tmp/apache
-sudo cp /etc/apache2/sites-available/default-ssl.conf /etc/apache2/sites-available/vhost-ssl.conf
-sudo sed 's/\/var\/www\/html/\/home\/pi\/Desktop\/gc/g' /etc/apache2/sites-available/vhost-ssl.conf |\
-    sudo tee /tmp/vhost && sudo cat /tmp/vhost |\
-    sudo tee /etc/apache2/sites-available/vhost-ssl.conf && sudo rm -f /tmp/vhost
-sudo sed 's/\/etc\/ssl\/certs\/ssl-cert-snakeoil\.pem/\/home\/pi\/_gc\/srv\/crt\/server\.crt/g' /etc/apache2/sites-available/vhost-ssl.conf |\
-    sudo tee /tmp/vhost && sudo cat /tmp/vhost |\
-    sudo tee /etc/apache2/sites-available/vhost-ssl.conf && sudo rm -f /tmp/vhost
-sudo sed 's/\/etc\/ssl\/private\/ssl-cert-snakeoil\.key/\/home\/pi\/_gc\/srv\/crt\/server\.key/g' /etc/apache2/sites-available/vhost-ssl.conf |\
-    sudo tee /tmp/vhost && sudo cat /tmp/vhost |\
-    sudo tee /etc/apache2/sites-available/vhost-ssl.conf && sudo rm -f /tmp/vhost
+
+# Apache設定
+if [ ! -f /etc/apache2/sites-available/000-default.conf.orig ]; then
+    sudo cp /etc/apache2/sites-available/000-default.conf /etc/apache2/sites-available/000-default.conf.orig
+    sudo sh -c 'cat << EOF > /etc/apache2/sites-available/000-default.conf
+<VirtualHost *:80>
+        ServerAdmin webmaster@localhost
+        DocumentRoot /home/pi/Desktop/gc
+
+        ErrorLog \${APACHE_LOG_DIR}/error.log
+        CustomLog \${APACHE_LOG_DIR}/access.log combined
+</VirtualHost>
+EOF'
+fi
+if [ ! -f //etc/apache2/apache2.conf.orig ]; then
+    sudo cp /etc/apache2/apache2.conf /etc/apache2/apache2.conf.orig
+    sudo sh -c 'cat << EOF > /etc/apache2/apache2.conf
+DefaultRuntimeDir \${APACHE_RUN_DIR}
+PidFile \${APACHE_PID_FILE}
+Timeout 300
+KeepAlive On
+MaxKeepAliveRequests 100
+KeepAliveTimeout 5
+
+User \${APACHE_RUN_USER}
+Group \${APACHE_RUN_GROUP}
+
+HostnameLookups Off
+
+ErrorLog \${APACHE_LOG_DIR}/error.log
+LogLevel warn
+
+IncludeOptional mods-enabled/*.load
+IncludeOptional mods-enabled/*.conf
+Include ports.conf
+
+<Directory />
+        Options FollowSymLinks
+        AllowOverride None
+        Require all denied
+</Directory>
+
+<Directory /usr/share>
+        AllowOverride None
+        Require all granted
+</Directory>
+
+<Directory /var/www/>
+        Options Indexes FollowSymLinks
+        AllowOverride None
+        Require all granted
+</Directory>
+
+<Directory /home/pi/Desktop/gc>
+        Options Indexes FollowSymLinks
+        AllowOverride None
+        Require all granted
+</Directory>
+
+AccessFileName .htaccess
+
+<FilesMatch "^\.ht">
+        Require all denied
+</FilesMatch>
+
+LogFormat "%v:%p %h %l %u %t \"%r\" %>s %O \"%{Referer}i\" \"%{User-Agent}i\"" vhost_combined
+LogFormat "%h %l %u %t \"%r\" %>s %O \"%{Referer}i\" \"%{User-Agent}i\"" combined
+LogFormat "%h %l %u %t \"%r\" %>s %O" common
+LogFormat "%{Referer}i -> %U" referer
+LogFormat "%{User-agent}i" agent
+
+IncludeOptional conf-enabled/*.conf
+IncludeOptional sites-enabled/*.conf
+EOF'
+fi
+
+sudo sh -c 'cat << EOF > /etc/apache2/sites-available/vhost-ssl.conf
+<IfModule mod_ssl.c>
+        <VirtualHost _default_:443>
+                ServerAdmin webmaster@localhost
+
+                DocumentRoot /home/pi/Desktop/gc
+
+                ErrorLog \${APACHE_LOG_DIR}/error.log
+                CustomLog \${APACHE_LOG_DIR}/access.log combined
+
+                SSLEngine on
+                SSLCertificateFile        /home/pi/_gc/srv/crt/server.crt
+                SSLCertificateKeyFile /home/pi/_gc/srv/crt/server.key
+
+                <FilesMatch "\.(cgi|shtml|phtml|php)$">
+                                SSLOptions +StdEnvVars
+                </FilesMatch>
+                <Directory /usr/lib/cgi-bin>
+                                SSLOptions +StdEnvVars
+                </Directory>
+        </VirtualHost>
+</IfModule>
+EOF'
+
 sudo a2ensite vhost-ssl
 sudo a2enmod ssl
 sudo systemctl restart apache2
-echo '@/usr/bin/chromium-browser https://localhost/top --enable-experimental-web-platform-features' | sudo tee -a /etc/xdg/lxsession/LXDE-pi/autostart
-sudo sed 's/Exec=\/usr\/bin\/x-www-browser\s%u/Exec=\/usr\/bin\/x-www-browser --enable-experimental-web-platform-features %u/g' /usr/share/raspi-ui-overrides/applications/lxde-x-www-browser.desktop |\
-    sudo tee /tmp/xbrowser && sudo cat /tmp/xbrowser |\
-    sudo tee /usr/share/raspi-ui-overrides/applications/lxde-x-www-browser.desktop && sudo rm -f /tmp/xbrowser
-sudo sed 's/Exec=chromium-browser/Exec=chromium-browser --enable-experimental-web-platform-features/g' /usr/share/applications/chromium-browser.desktop |\
-    sudo tee /tmp/chbrowser && sudo cat /tmp/chbrowser |\
-    sudo tee /usr/share/applications/chromium-browser.desktop && sudo rm -f /tmp/chbrowser
+grep '--enable-experimental-web-platform-features' /usr/share/raspi-ui-overrides/applications/lxde-x-www-browser.desktop
+if [ $? = 1 ]; then
+    sudo sed 's/Exec=\/usr\/bin\/x-www-browser\s%u/Exec=\/usr\/bin\/x-www-browser --enable-experimental-web-platform-features %u/g' /usr/share/raspi-ui-overrides/applications/lxde-x-www-browser.desktop |\
+        sudo tee /tmp/xbrowser && sudo cat /tmp/xbrowser |\
+        sudo tee /usr/share/raspi-ui-overrides/applications/lxde-x-www-browser.desktop && sudo rm -f /tmp/xbrowser
+fi
+grep '--enable-experimental-web-platform-features' /usr/share/applications/chromium-browser.desktop
+if [ $? = 1 ]; then
+    sudo sed 's/Exec=chromium-browser/Exec=chromium-browser --enable-experimental-web-platform-features/g' /usr/share/applications/chromium-browser.desktop |\
+        sudo tee /tmp/chbrowser && sudo cat /tmp/chbrowser |\
+        sudo tee /usr/share/applications/chromium-browser.desktop && sudo rm -f /tmp/chbrowser
+fi
 
 # 証明書追加
 certfile="/home/pi/_gc/srv/crt/ca.crt"
@@ -165,17 +288,19 @@ done
 # Arduino IDE 追加
 cd /home/pi/
 mkdir /home/pi/Applications/
-tar xvf arduino-1.8.6-linuxarm.tar.xz
-mv arduino-1.8.6 /home/pi/Applications/
+if [ ! -d /home/pi/Applications/arduino-1.8.13/ ]; then
+    tar xvf arduino-1.8.13-linuxarm.tar.xz
+    mv arduino-1.8.13 /home/pi/Applications/
+fi
 cd /home/pi/Applications/
-ln -s arduino-1.8.6 arduino
+ln -s arduino-1.8.13 arduino
 cd /home/pi/Applications/arduino/
 ./install.sh
-rm -f /home/pi/arduino-1.8.6-linuxarm.tar.xz
+rm -f /home/pi/arduino-1.8.13-linuxarm.tar.xz
 cd /home/pi/
 
 # upgradeを保留を解除
-echo raspberrypi-ui-mods install | sudo dpkg --set-selections
+sudo apt-mark auto raspberrypi-ui-mods
 # 上をアップグレード
 sudo apt-get -y upgrade
